feat: Сдлеать модуль "Бронирования" без валидации

+++ src/reservation/reservation.service.ts
@@ -1,66 +1,62 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { ICreateReservation, IReservationSearchOptions } from './dto';
+import { IReservationSearchOptions } from './dto';
 import { IDocumentReservation, Reservation } from './model';
 import { TID } from 'src/types';
+import { HotelRoomService } from 'src/hotel/hotel-room/hotel-room.service';
 
 @Injectable()
 export class ReservationService {
   constructor(
     @InjectModel(Reservation.name)
     private ReservationModel: Model<IDocumentReservation>,
+    private readonly hotelRoomService: HotelRoomService,
   ) {}
 
-  async addReservation(
-    data: ICreateReservation,
-  ): Promise<IDocumentReservation | null> {
-    try {
-      const filter: IReservationSearchOptions = {
-        userId: data.userId,
-        dateStart: data.dateStart,
-        dateEnd: data.dateEnd,
-      };
-      const reservations: Array<IDocumentReservation | null> =
-        await this.getReservations(filter);
-
-      if (reservations.length === 0) {
-        throw new Error('Нет доступных мест для бронирования');
-      }
-
-      const newReservation: IDocumentReservation = new this.ReservationModel(
-        data,
-      );
-      return newReservation.save();
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        console.log(error.message);
-      }
-      return null;
+  async addReservation(data): Promise<IDocumentReservation> {
+    const hotelRoom = await this.hotelRoomService.findById(data.roomId);
+    if (!hotelRoom) {
+      throw new BadRequestException('Нет такого номера');
     }
+
+    if (!hotelRoom.isEnabled) {
+      throw new BadRequestException('Данный номер отключен');
+    }
+
+    const reservation = {
+      ...data,
+      hotelId: hotelRoom.hotel,
+    };
+
+    const newReservation: IDocumentReservation = new this.ReservationModel(
+      reservation,
+    );
+    return newReservation.save();
   }
 
   async getReservations(
     filter: IReservationSearchOptions,
-  ): Promise<Array<IDocumentReservation> | null> {
-    try {
-      return this.ReservationModel.find(filter).select('-__v');
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        console.log(error.message);
-      }
-      return null;
-    }
+    select?: string,
+    populate?: Array<{ path: string; select: string }>,
+  ): Promise<Array<IDocumentReservation>> {
+    return this.ReservationModel.find(filter)
+      .select('-__v' + select ? ' ' + select : '')
+      .populate(populate);
   }
 
-  async removeReservation(id: TID) {
-    try {
-      return this.ReservationModel.findByIdAndDelete(id);
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        console.log(error.message);
-      }
-      return false;
+  async removeReservation(id: TID, userId?: string) {
+    const reservation = await this.ReservationModel.findById(id);
+    if (!reservation) {
+      throw new BadRequestException('Данной бронь отсутствует');
     }
+
+    if (userId && reservation.userId.toString() !== userId.toString()) {
+      throw new BadRequestException(
+        'Данная бронь не принадлежит текущему пользователю',
+      );
+    }
+
+    return this.ReservationModel.findByIdAndDelete(id);
   }
 }
